Extract periodic timers concern
